Split isolated-feed tests into NugetServiceFeedTests; guard NUGET_PACKAGES
- Move the private nuget.config (isolated feed) fixtures and configuration/
  packageSourceMapping/local-feed tests out of NugetServiceTests.cs into a new
  NugetServiceFeedTests class so each file stays under the repo's ~1,000-line
  limit and has one focused responsibility (core/live-integration vs. isolated
  feed/config). NugetServiceTests.cs drops from ~1,050 to ~470 lines.
- Guard the two local-feed install tests with Assert.Inconclusive when
  NUGET_PACKAGES is set: that variable overrides the config globalPackagesFolder,
  so the install would hit the shared cache and could pass without exercising the
  local feed. Mirrors the existing guard on GetNuGetGlobalPackagesDir test.
